refactor(node): move _fixed_queue.ts to node/internal
