+++ node/_next_tick.ts
@@ -6,7 +6,7 @@
 import { core } from "./_core.ts";
 import { validateCallback } from "./internal/validators.mjs";
 import { _exiting } from "./_process/exiting.ts";
-import { FixedQueue } from "./_fixed_queue.ts";
+import { FixedQueue } from "./internal/fixed_queue.ts";
 
 interface Tock {
   callback: (...args: Array<unknown>) => void;
